Port staking tests to typescript (#55)
* Fix up the Distributor tests

Also moves them to typescript
* Ports Staking test over to typescript
Had to edit the OLDwsOHM.sol contract to rename the IsOHM interface to
IsOHMOLD because it was conflicting with the new IsOHM interface. Only
the old type definitions were being created.
